+++ src/app/pages/billing-main/billing-main.module.ts
@@ -10,12 +10,14 @@ import { InfoCardComponent } from '@app/shared/components/info-card/info-card.co
 import { MessageNoDataComponent } from '@app/shared/components/message-no-data/message-no-data.component';
 import { BillingRoutingModule } from './billing-routing.module';
 import { NumberFormatInputPipe } from '@app/shared/pipes/numberFormatInput.pipe';
+import { PreviousBillingComponent } from './previous-billing/previous-billing.component';
 
 
 @NgModule({
   declarations: [
     BillingComponent,
-    NumberFormatInputPipe
+    NumberFormatInputPipe,
+    PreviousBillingComponent
   ],
   imports: [
     CommonModule,
