+++ src/app/pages/billing-main/billing/billing.component.scss
@@ -1,13 +1,5 @@
 .c-billing {
-    .table-content {
-        th:nth-child(1) {
-            padding-top: 7px !important;
-            padding-left: 0 !important;
-        }
+    .c-scroll-table {
+        height: calc(100vh - 310px)!important;
     }
-}
-
-.btns-excel-report {
-    display: flex;
-    justify-content: end;
 }
