+++ src/app/pages/billing-main/billing/billing.component.ts
@@ -11,15 +11,13 @@ import { BillingService } from '../billing.service';
 import { selectPageIndex, selectPageSize } from '@app/core/store/selectors/paginator.selector';
 import { FormControl } from '@angular/forms';
 import { updatePagination } from '@app/core/store/actions/paginator.actions';
-import { FilterState, GeneralFilters, notificationData, UserInfo } from '@app/shared/models/general-models';
+import { FilterState, GeneralFilters, UserInfo } from '@app/shared/models/general-models';
 import { PeriodicElement } from '@app/pages/plants-main/plants-model';
 import { SelectionModel } from '@angular/cdk/collections';
 import { EncryptionService } from '@app/shared/services/encryption.service';
 import { NotificationDataService } from '@app/shared/services/notificationData.service';
 import { NOTIFICATION_CONSTANTS } from '@app/core/constants/notification-constants';
 import { MatDialog } from '@angular/material/dialog';
-import { NotificationComponent } from '@app/shared/components/notification/notification.component';
-import { HttpErrorResponse } from '@angular/common/http';
 
 @Component({
   selector: 'app-billing',
@@ -42,18 +40,21 @@ export class BillingComponent implements OnDestroy, OnInit, AfterViewChecked, Af
   pageIndex: number = 1;
   totalItems: number = 0;
   displayedColumns: string[] = [
-    'checkbox',
-    'rpu',
+    'clientId',
+    'subClientId',
     'clientName',
-    'status',
-    'plantName',
-    'year',
+    'subClient',
+    'rfc',
     'month',
-    'energyGeneration',
+    'year',
+    'billingPeriodStart',
+    'billingPeriodEnd',
+    'receiptGenerationDate',
+    'cfcContact',
     'rate',
-    'amount',
-    'amountWithIva',
-    'actions',
+    'production',
+    'previousPaymentAmount',
+    'totalAmount'
   ];
   pageSizeSub!: Subscription;
   pageIndexSub!: Subscription;
@@ -63,7 +64,212 @@ export class BillingComponent implements OnDestroy, OnInit, AfterViewChecked, Af
 
   formatTimer: any;
 
-  modifiedElements: any[] = [];
+  dummy: any[] = [
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+    {
+      clientId: '001',
+      subClientId: '026',
+      clientName: 'Customer name',
+      subClient: 'Plant name',
+      rfc: 'ER02856969UX3',
+      month: 'Jan',
+      year: '2025',
+      billingPeriodStart: '01/01/2025',
+      billingPeriodEnd: '31/01/2025',
+      receiptGenerationDate: '31/01/2025',
+      cfcContact: 'CFE-987654',
+      rate: '$000,000.00 MXN',
+      production: '5,000 kWh',
+      previousPaymentAmount: '$000,000.00 MXN',
+      totalAmount: '$000,000.00 MXN',
+    },
+  ];
 
   searchBar = new FormControl('');
 
@@ -124,204 +330,27 @@ export class BillingComponent implements OnDestroy, OnInit, AfterViewChecked, Af
   }
 
   getBilling(searchTerm: string = '') {
-    const filters = {
-      plantName: searchTerm,
-      pageSize: this.pageSize,
-      page: this.pageIndex,
-      ...this.generalFilters
-    };
+    // const filters = {
+    //   plantName: searchTerm,
+    //   pageSize: this.pageSize,
+    //   page: this.pageIndex,
+    //   ...this.generalFilters
+    // };
 
-    this.moduleServices.getBillingData(filters).subscribe({
-      next: (response: entity.DataBillingTableMapper) => {
-        this.dataSource.data = response?.data;
-        this.totalItems = response?.totalItems;
-        this.dataSource.sort = this.sort;
-        this.pageIndex = filters.page;
-      },
-      error: error => {
-        this.notificationService.notificacion(`Talk to the administrator.`, 'alert');
-        console.log(error);
-      }
-    });
-  }
+    // this.moduleServices.getBillingData(filters).subscribe({
+    //   next: (response: entity.DataBillingTableMapper) => {
+    //     this.dataSource.data = response?.data;
+    //     this.totalItems = response?.totalItems;
+    //     this.dataSource.sort = this.sort;
+    //     this.pageIndex = filters.page;
+    //   },
+    //   error: error => {
+    //     this.notificationService.notificacion(`Talk to the administrator.`, 'alert');
+    //     console.log(error);
+    //   }
+    // });
+    this.dataSource.data = this.dummy
 
-  generateInvoice() {
-    const objData: entity.CreateInvoice | any = {
-      clientId: this.userInfo.clientes[0],
-      ...this.generalFilters
-    }
-
-    this.moduleServices.generateInvoice(objData).subscribe({
-      next: (response: any) => {
-        this.notificationService.notificacion(`Energy generation has been updated.`, 'save')
-        this.getBilling(this.searchBar?.value!);
-      },
-      error: (error: HttpErrorResponse) => {
-        const errorMessages = error?.error?.errors?.errors.map((e: any) => e.descripcion)
-        this.modalErrors(errorMessages);
-      }
-    });
-  }
-
-  generateInvoiceAction() {
-    this.createNotificationModal(true);
-  }
-
-  confirmInvoiceAction(oneInvoice?: entity.DataBillingTable) {
-    if (oneInvoice) this.oneConfirmInvoice = oneInvoice;
-    this.createNotificationModal(false);
-  }
-
-  updateMultipleInvoiceStatuses() {
-    const objData: entity.DataBillingTable | any[] = this.selection?.selected.length ? this.selection?.selected : [this.oneConfirmInvoice]
-    const filteredInvoices = objData.filter(invoice => invoice.status != 2);
-
-    filteredInvoices.forEach((data: any) => {
-      delete data.amountWithIva;
-      delete data.formattedGeneratedEnergyKwh;
-      delete data.originalGeneratedEnergyKwh;
-      delete data.formattedAmount;
-      delete data.formattedAmountWithIva;
-      delete data.formattedRate;
-      delete data.formatterStatus;
-    });
-
-    this.moduleServices.updateMultipleInvoiceStatuses(filteredInvoices).subscribe({
-      next: (_) => {
-        this.notificationService.notificacion(`The changes have been successfully saved, and the presented data has been adjusted.`, 'save')
-          .afterClosed()
-          .subscribe((_ => {
-            this.getBilling(this.searchBar?.value!)
-            this.modifiedElements = []
-          }));
-      },
-      error: (error: HttpErrorResponse) => {
-        const errorMessages = error?.error?.errors?.errors.map((e: any) => e.descripcion)
-        this.modalErrors(errorMessages);
-      }
-    });
-  }
-
-  updateModifiedElements(oneInvoice?: entity.DataBillingTable) {
-    let invoices: entity.DataBillingTable | any[] = oneInvoice ? [oneInvoice] : this.modifiedElements
-
-    const filteredInvoices = invoices.map((data) => ({
-      generatedEnergyKwh: data.generatedEnergyKwh,
-      subtotal: data.subtotal,
-      iva: data.iva,
-      total: data.total,
-      status: data.status,
-      invoiceId: data.invoiceId,
-    }));
-
-    this.moduleServices.saveBillingTableData(filteredInvoices).subscribe({
-      next: () => {
-        this.notificationService.notificacion(`The changes have been successfully saved, and the presented data has been adjusted.`, 'save')
-          .afterClosed()
-          .subscribe((_ => {
-            this.getBilling(this.searchBar?.value!)
-            this.modifiedElements = []
-          }));
-      },
-      error: (error: HttpErrorResponse) => {
-        const errorMessages = error?.error?.errors?.errors.map((e: any) => e.descripcion)
-        this.modalErrors(errorMessages);
-      }
-    });
-  }
-
-  createNotificationModal(generate: boolean) {
-    let dataNotificationModal: notificationData
-    if (generate) dataNotificationModal = this.notificationDataService.invoicesNotificationData(generate, this.generalFilters)!;
-    else dataNotificationModal = this.notificationDataService.invoicesNotificationData(generate)!;
-
-    const dialogRef = this.dialog.open(NotificationComponent, {
-      width: '540px',
-      data: dataNotificationModal
-    })
-    dialogRef.afterClosed().subscribe(result => {
-      if (result && result.confirmed) {
-        if (generate) this.generateInvoice()
-        else this.updateMultipleInvoiceStatuses()
-      }
-    });
-  }
-
-  handleInput(event: any, element: any, isBlurEvent: boolean = false) {
-    const cleanedValue: any = this.cleanFormattedValue(event.target.value);
-
-    element.generatedEnergyKwh = cleanedValue;
-    element.formattedGeneratedEnergyKwh = this.getFormattedValue(cleanedValue);
-
-    if (isBlurEvent) event.target.value = element.formattedGeneratedEnergyKwh;
-    this.trackChanges(element);
-  }
-
-  trackChanges(element: any) {
-    const index = this.modifiedElements.findIndex(el => el.invoiceId === element.invoiceId);
-
-    const cleanedOriginalEnergy = this.cleanFormattedValue(element.originalGeneratedEnergyKwh);
-    const cleanedCurrentEnergy = this.cleanFormattedValue(element.generatedEnergyKwh);
-
-    const isEnergyChanged = cleanedOriginalEnergy !== cleanedCurrentEnergy;
-    const isEnergyCleared = cleanedCurrentEnergy === null;
-
-    if (isEnergyCleared && index !== -1) this.modifiedElements.splice(index, 1);
-    else if (isEnergyChanged && index === -1) this.modifiedElements.push({ ...element });
-    else if (!isEnergyChanged && index !== -1) this.modifiedElements.splice(index, 1);
-    else if (isEnergyChanged && index >= 0) this.modifiedElements[index] = { ...element };
-  }
-
-  cleanFormattedValue(value: any): number | null {
-    if (!value) return null;
-    const cleanedValue = value.toString().replace(/[^\d.-]/g, '');
-    const parsedValue = parseFloat(cleanedValue);
-    return isNaN(parsedValue) ? null : parsedValue;
-  }
-
-  getFormattedValue(value: string): string {
-    const numberValue = parseFloat(value);
-    return !isNaN(numberValue)
-      ? numberValue.toLocaleString('en-US', {
-        minimumFractionDigits: 2,
-        maximumFractionDigits: 2
-      })
-      : '';
-  }
-
-  navigate(link: string) {
-    this.router.navigateByUrl(link);
-  }
-
-  isAllSelected() {
-    const numSelected = this.selection.selected.length;
-    const numRows = this.dataSource.data.length;
-    return numSelected === numRows;
-  }
-
-  toggleAllRows() {
-    if (this.isAllSelected()) {
-      this.selection.clear();
-
-      return;
-    }
-
-    this.selection.select(...this.dataSource.data);
-  }
-
-  toggleRow(row: any) {
-    this.selection.toggle(row);
-  }
-
-  checkboxLabel(row?: PeriodicElement): string {
-    if (!row) return `${this.isAllSelected() ? 'deselect' : 'select'} all`;
-    return `${this.selection.isSelected(row) ? 'deselect' : 'select'} row ${row.position + 1}`;
-  }
-
-  getServerData(event: PageEvent): void {
-    if (event.pageSize !== this.pageSize || event.pageIndex !== this.pageIndex - 1) {
-      this.store.dispatch(updatePagination({ pageIndex: event.pageIndex, pageSize: event.pageSize }));
-    }
   }
 
   changePageSize(event: any) {
@@ -336,34 +365,10 @@ export class BillingComponent implements OnDestroy, OnInit, AfterViewChecked, Af
     this.getBilling();
   }
 
-  exportInformation() {
-    this.moduleServices.downloadExcelReport({
-      startDate: this.generalFilters.startDate,
-      endDate: this.generalFilters.endDate!,
-      plantName: this.searchBar?.value!
-    }).subscribe({
-      next: (response: Blob) => {
-        const blob = new Blob([response], { type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet' });
-        const link = document.createElement('a');
-        const url = window.URL.createObjectURL(blob);
-        link.href = url;
-        link.download = 'ExcelTemplate.xlsx';
-        link.click();
-        window.URL.revokeObjectURL(url);
-      },
-      error: error => {
-        this.notificationService.notificacion('Talk to the administrator.', 'alert');
-        console.log(error);
-      }
-    })
-  }
-
-  modalErrors(errors: any) {
-    const dataNotificationModal: notificationData = this.notificationDataService.errors(errors)!;
-    this.dialog.open(NotificationComponent, {
-      width: '540px',
-      data: dataNotificationModal
-    })
+  getServerData(event: PageEvent): void {
+    if (event.pageSize !== this.pageSize || event.pageIndex !== this.pageIndex - 1) {
+      this.store.dispatch(updatePagination({ pageIndex: event.pageIndex, pageSize: event.pageSize }));
+    }
   }
 
   ngOnDestroy(): void {
