+++ src/app/pages/rates-main/rates/rates.component.html
@@ -1,5 +1,5 @@
 <div class="c-main px-16">
-    <div class="grid md:grid-cols-3 sm:grid-cols-2 gap-x-6 gap-y-4 md:gap-5">
+    <div class="grid md:grid-cols-2 sm:grid-cols-2 gap-x-6 gap-y-4 md:gap-5">
         <div class="col-span-1">
             <div class="btns-excel">
                 <button (click)="downloadExcel()" class="btn-action-add">
@@ -16,7 +16,7 @@
             </div>
         </div>
 
-        <div class="col-span-1 d-flex justify-center">
+        <div class="col-span-1 flex justify-end">
             <div class="search-add">
                 <form class="relative w-80 ms-auto c-search">
                     <label class="input-search">
@@ -28,7 +28,7 @@
             </div>
         </div>
 
-        <div class="col-span-1"></div>
+        <!-- <div class="col-span-1"></div> -->
     </div>
 
     <div class="separator-cards mt-4 mb-4"></div>
